test: 삭제된 게시글을 다시 삭제했을 떄 발생하는 예외 추가 및 테스트 코드 작성

+++ week4/src/main/java/com/kyle/week4/service/PostService.java
@@ -132,6 +132,10 @@ public class PostService {
             throw new CustomException(PERMISSION_DENIED);
         }
 
+        if (post.isDeleted()) {
+            throw new CustomException(ALREADY_DELETED_POST);
+        }
+
         post.delete();
     }
